requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117

torch==1.13.1+cu117 
torchvision==0.14.1+cu117
faiss-cpu==1.7.3
scikit-learn==1.2.0
scipy==1.10.0
transformers==4.25.1
timm==0.6.12
torchmetrics==0.11.1
einops
tqdm